refactor: modularize event handler tests to respect 300-line limit
